Refactor game type folder structure
