+++ src/app/features/gameTypes/index.ts
@@ -1,6 +1,6 @@
 import type { GameTypeId } from "../../../api/services/game/types";
-import { reactVersus } from "./versus/react/gameType";
-import { excaliburVersus } from "./versus/excalibur/gameType";
+import { reactVersus } from "./versus/gameType";
+import { excaliburVersus } from "./excalibur/gameType";
 import type { GameType } from "./GameType";
 
 export const gameTypeList: GameType[] = [reactVersus, excaliburVersus];
