+++ src/app/features/gameTypes/versus/components/Card.tsx
