+++ src/app/features/gameTypes/versus/scenes/DeckSelectScene.tsx
@@ -2,10 +2,10 @@ import Button from "@mui/material/Button";
 import { useState } from "react";
 import Stack from "@mui/material/Stack";
 import Paper from "@mui/material/Paper";
-import { adapter } from "../../runtimeDefinition";
-import { Center } from "../../../../../components/Center";
-import type { RuntimeDeck, RuntimeGenerics } from "../../../../compiler/types";
-import { Select } from "../../../../../controls/Select";
+import { adapter } from "../runtimeDefinition";
+import { Center } from "../../../../components/Center";
+import type { RuntimeDeck, RuntimeGenerics } from "../../../compiler/types";
+import { Select } from "../../../../controls/Select";
 
 export function DeckSelectScene() {
   const { startBattle } = adapter.useRuntimeActions();
