+++ src/app/features/gameTypes/versus/scenes/index.tsx
@@ -1,4 +1,4 @@
-import { adapter } from "../../runtimeDefinition";
+import { adapter } from "../runtimeDefinition";
 import { DeckSelectScene } from "./DeckSelectScene";
 import { BattleScene } from "./BattleScene";
 import { ResultScene } from "./ResultScene";
